Added review app commands: enable, disable
